features/requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
scikit-learn
requests
python-dotenv